requirements.txt
fastapi>=0.112.0,<0.113.0
uvicorn>=0.30.0
httpx>=0.27.0
pydantic>=2.7.0,<3.0.0
pyyaml>=6
loguru
